// Frontend/QuanLyCuTru/src/app/models/standards/nguoidung.ts
export interface NguoiDung {
  id: number;
  hoTen: String;
  gioiTinh: boolean;
  sinhNhat: Date;
  soNha: String;
  duong: String;
  phuong: String;
  quan: String;
  thanhPho: String;
  noiSinh: String;
  queQuan: String;
  quocTich: String;
  avatar: String;
  identityId: String;
  chucVuId: number;
}